refactor: define VisualTransformation as a class
